DBManager/requirements.txt
alembic==1.11.1
async-timeout==4.0.2
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
everapi==0.1.1
Flask==2.3.2
Flask-Migrate==4.0.4
flask-redis==0.4.0
flask-sqlalchemy==3.0.5
freecurrencyapi==0.1.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
typing-extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.16.2
